rename and move kafka consumer service
