+++ passenger-service/src/main/java/io/voitovich/yura/passengerservice/event/service/KafkaRatingConsumerService.java
@@ -1,8 +1,8 @@
-package io.voitovich.yura.passengerservice.event;
+package io.voitovich.yura.passengerservice.event.service;
 
 import io.voitovich.yura.passengerservice.event.model.ReceiveRatingModel;
 
-public interface KafkaConsumerService {
+public interface KafkaRatingConsumerService {
 
     void consumeRating(ReceiveRatingModel model);
 }
