+++ passenger-service/src/main/java/io/voitovich/yura/passengerservice/event/service/impl/KafkaRatingConsumerServiceImpl.java
@@ -1,6 +1,6 @@
-package io.voitovich.yura.passengerservice.event.impl;
+package io.voitovich.yura.passengerservice.event.service.impl;
 
-import io.voitovich.yura.passengerservice.event.KafkaConsumerService;
+import io.voitovich.yura.passengerservice.event.service.KafkaRatingConsumerService;
 import io.voitovich.yura.passengerservice.event.model.ReceiveRatingModel;
 import io.voitovich.yura.passengerservice.service.RatingService;
 import lombok.extern.slf4j.Slf4j;
@@ -8,13 +8,13 @@ import org.springframework.stereotype.Service;
 
 @Slf4j
 @Service
-public class KafkaConsumerServiceImpl implements KafkaConsumerService {
+public class KafkaRatingConsumerServiceImpl implements KafkaRatingConsumerService {
 
     private final String topicName = "rate_passenger_topic";
 
     private final RatingService ratingService;
 
-    public KafkaConsumerServiceImpl(RatingService ratingService) {
+    public KafkaRatingConsumerServiceImpl(RatingService ratingService) {
         this.ratingService = ratingService;
     }
